move : 상수 파일 이동으로 import문 수정
